Added visible parameter to define bones that are only visible
There had been a discussion in the past about this thing in ViUR 2.x and their related ViUR forms already, but I'm still of the opinion that it's to developer's choice which bones are necessary and should be asked for. The core will check the rest when something is insufficient. Bones can also be pre-filled by addSkel or similar.