Improved @html5.tag decorator, "flare-"-prefix
This improves the @html5.tag-decorator so that it can either be called with a tag-name to associate the following class with, or without, so the class-name is used as tag-name.

Additionally added "flare-"-prefix to all flare-specific components. This is a breaking change, it requires update the HTML-code inside existing projects.