Update packages.json for user-specified packages (#44)
Line 146 of `get-pyodide.py` overwrites the user-specified `packages` variable and line 149 only keeps the default packages, ignoring the user-specified ones. The result is that user-specified packages are fetched, but are not importable.

Extend the default global `PACKAGES` variable with the user-specified packages and use this variable when referencing packages to fetch and include.