Improving flare.py (#46)
* Improving and simplifying flare.py

* Minify without using os.system()

os.system() is ugly here, as we are in a Python ecosystem we should also run tools written in Python direcly. There are plenty of other issues in here that shal be solved.

* Removing cleanString

In case an invalid path was provided to PROJECT_WORKSPACE, the script shall raise an exception. It's on the user's responsibility to provide correct pathes here.

* Ok stuff from today ey enough man

* Substituting all os.system() calls by Python library calls.

* Bugfix for handling flare/flare in zipfiles